+++ src/absim_atmega32u4.hpp
@@ -89,7 +89,7 @@ ARDENS_FORCEINLINE uint32_t atmega32u4_t::advance_cycle()
         t = std::min<uint64_t>(t, timer4.next_update_cycle - cycle_count);
         t = std::min<uint64_t>(t, usb_next_update_cycle - cycle_count);
 
-        max_merged_cycles = std::min<uint64_t>(t, 1024) - MAX_INSTR_CYCLES;
+        max_merged_cycles = (uint32_t)std::min<uint64_t>(t, 1024) - MAX_INSTR_CYCLES;
 
         single_instr_only = (
             spi_busy ||
